Moving Enrique and SignatureStamp out of characters.  Enrique is now DataSource; SignatureStamp now lives in signature.
