+++ nkms/crypto/signature.py
@@ -1,6 +1,8 @@
 from nkms.crypto import api as API
 from umbral.keys import UmbralPublicKey
 from cryptography.hazmat.primitives.asymmetric.utils import decode_dss_signature, encode_dss_signature
+from nkms.crypto.api import keccak_digest
+from bytestring_splitter import BytestringSplitter
 
 
 class Signature(object):
@@ -59,3 +61,60 @@ class Signature(object):
     def __eq__(self, other):
         # TODO: Consider constant time
         return bytes(self) == bytes(other) or self._der_encoded_bytes() == other
+
+
+signature_splitter = BytestringSplitter(Signature)
+
+
+class SignatureStamp(object):
+    """
+    Can be called to sign something or used to express the signing public
+    key as bytes.
+    """
+
+    def __init__(self, character):
+        self.character = character
+
+    def __call__(self, *args, **kwargs):
+        return self.character.sign(*args, **kwargs)
+
+    def __bytes__(self):
+        from nkms.crypto.powers import SigningPower
+        return bytes(self.character.public_key(SigningPower))
+
+    def __hash__(self):
+        return int.from_bytes(self, byteorder="big")
+
+    def __eq__(self, other):
+        return other == bytes(self)
+
+    def __add__(self, other):
+        return bytes(self) + other
+
+    def __radd__(self, other):
+        return other + bytes(self)
+
+    def __len__(self):
+        return len(bytes(self))
+
+    def as_umbral_pubkey(self):
+        from nkms.crypto.powers import SigningPower
+        return self.character.public_key(SigningPower)
+
+    def fingerprint(self):
+        """
+        Hashes the key using keccak-256 and returns the hexdigest in bytes.
+
+        :return: Hexdigest fingerprint of key (keccak-256) in bytes
+        """
+        return keccak_digest(bytes(self)).hex().encode()
+
+
+class StrangerStamp(SignatureStamp):
+    """
+    SignatureStamp of a stranger (ie, can only be used to glean public key, not to sign)
+    """
+
+    def __call__(self, *args, **kwargs):
+        message = "This isn't your SignatureStamp; it belongs to {} (a Stranger).  You can't sign with it."
+        raise TypeError(message.format(self.character))
